pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "enterprise-rag-production"
version = "0.1.0"
description = "Production-grade RAG pipeline reference implementation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Sudarshan Angirash", email = "sudaangi@techtonicis.com"}
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "openai>=1.40.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.9",
    "pinecone-client>=3.2.0",
    "weaviate-client>=4.7.0",
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.3",
    "tiktoken>=0.7.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "prometheus-client>=0.20.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
